Fix various floating point precision issues

+++ src/Cocoa/MyOpenGLView.m
@@ -738,7 +738,7 @@ FAIL:
 {
 	double mx = [theEvent locationInWindow].x - viewSize.width/2.0;
 	double my = [theEvent locationInWindow].y - viewSize.height/2.0;
-		
+
 	if (display_z > 640.0)
 	{
 		mx /= viewSize.width * MAIN_GUI_PIXEL_WIDTH / display_z;
@@ -1257,13 +1257,13 @@ static void UnapplyCursorState(OOMouseInteractionMode mode)
 
 - (void) setFov:(float)value fromFraction:(BOOL)fromFraction
 {
-	_fov = fromFraction ? value : tan((value / 2) * M_PI / 180);
+	_fov = fromFraction ? value : tanf((value / 2.0f) * M_PI / 180.0f);
 }
 
 
 - (float) fov:(BOOL)inFraction
 {
-	return inFraction ? _fov : 2 * atan(_fov) * 180 / M_PI;
+	return inFraction ? _fov : 2 * atanf(_fov) * 180.0f / M_PI;
 }
 
 @end
